Meeting 2020-11-15
Meeting 2020-11-15

Talked about how when you're lifting loops to upstairs be mindful of the "lollipop" option where once lifted you get upstairs |__|___|_|__| and the geodesic you want is the ____ part.

At the outset for cube complexes you can have a hyperplane with several dimensions present and it could intersect itself. At this level fo generality hyperplanes are abstract cube complexes with maps into a cube complex that may not be embeddings. Once you have two-sidedness (think: 1-sidedness is like trying to define a push map from the core of the mobius band to the outside) these maps are embeddings and therefore you can think of a hyperplane as a literal subspace.

Even in the 2-sided scenario you can have things like circle 2x annulus 2x circle where the hyperplane is a circle and it is two sided but it gets wrapped at the edges. In the cover everything unwraps and you get homeomorphisms between lines

You can talk about things like characteristic maps for a cell complex and similarly for a cube complex or even the closed product neighborhood in the 2-sided case. These are maps from h x [0,1] that are injective on interior to C the whole cube complex. 

When we are proving that our extended core C is simply connected, we're sort of doing the process for making a cell complex but in reverse. i.e. we are starting with a space and giving a decomposition.